+++ src/main/java/eu/bausov/projects/srvpumpselection/web/security/AuthenticationFilter.java
@@ -26,6 +26,4 @@ public class AuthenticationFilter extends AbstractPreAuthenticatedProcessingFilt
     protected Object getPreAuthenticatedCredentials(HttpServletRequest httpServletRequest) {
         return "N/A";
     }
-
-
 }
